Allow gifs to play once without repetition
Adapted from http://stackoverflow.com/questions/24004363/convert-images-to-gif - no application extension is to be written if no "looping" in the .gif is desired. 

For legacy purposes, loops=0 is still considered an "infinite" loop, but now calling writeGif() with repeat=False works properly (instead of repeating once, as before).

Also, fixed some (but not all) typos in the documentation.